Remove NearestNeighbors class

+++ src/simplex.cpp
@@ -18,7 +18,7 @@ void simplex(TimeSeries &prediction, const TimeSeries &target, const LUT &lut)
         "lookup", distances.extent(0), KOKKOS_LAMBDA(int i) {
             float pred = 0.0f;
 
-            for (int j = 0u; j < distances.extent(1); j++) {
+            for (size_t j = 0; j < distances.extent(1); j++) {
                 pred += target(indices(i, j)) * distances(i, j);
             }
